+++ scheduler/system_sched_test.go
@@ -91,7 +91,7 @@ func TestSystemeSched_JobRegister_StickyAllocs(t *testing.T) {
 
 	// Create a job
 	job := mock.SystemJob()
-	job.TaskGroups[0].LocalDisk.Sticky = true
+	job.TaskGroups[0].EphemeralDisk.Sticky = true
 	noErr(t, h.State.UpsertJob(h.NextIndex(), job))
 
 	// Create a mock evaluation to register the job
@@ -150,7 +150,7 @@ func TestSystemeSched_JobRegister_StickyAllocs(t *testing.T) {
 	}
 }
 
-func TestSystemSched_JobRegister_LocalDiskConstraint(t *testing.T) {
+func TestSystemSched_JobRegister_EphemeralDiskConstraint(t *testing.T) {
 	h := NewHarness(t)
 
 	// Create a nodes
@@ -159,13 +159,13 @@ func TestSystemSched_JobRegister_LocalDiskConstraint(t *testing.T) {
 
 	// Create a job
 	job := mock.SystemJob()
-	job.TaskGroups[0].LocalDisk.DiskMB = 60 * 1024
+	job.TaskGroups[0].EphemeralDisk.SizeMB = 60 * 1024
 	noErr(t, h.State.UpsertJob(h.NextIndex(), job))
 
 	// Create another job with a lot of disk resource ask so that it doesn't fit
 	// the node
 	job1 := mock.SystemJob()
-	job1.TaskGroups[0].LocalDisk.DiskMB = 60 * 1024
+	job1.TaskGroups[0].EphemeralDisk.SizeMB = 60 * 1024
 	noErr(t, h.State.UpsertJob(h.NextIndex(), job1))
 
 	// Create a mock evaluation to register the job
@@ -1227,4 +1227,91 @@ func TestSystemSched_ChainedAlloc(t *testing.T) {
 	if len(newAllocs) != 2 {
 		t.Fatalf("expected: %v, actual: %v", 2, len(newAllocs))
 	}
+}
+
+func TestSystemSched_PlanWithDrainedNode(t *testing.T) {
+	h := NewHarness(t)
+
+	// Register two nodes with two different classes
+	node := mock.Node()
+	node.NodeClass = "green"
+	node.Drain = true
+	node.ComputeClass()
+	noErr(t, h.State.UpsertNode(h.NextIndex(), node))
+
+	node2 := mock.Node()
+	node2.NodeClass = "blue"
+	node2.ComputeClass()
+	noErr(t, h.State.UpsertNode(h.NextIndex(), node2))
+
+	// Create a Job with two task groups, each constrianed on node class
+	job := mock.SystemJob()
+	tg1 := job.TaskGroups[0]
+	tg1.Constraints = append(tg1.Constraints,
+		&structs.Constraint{
+			LTarget: "${node.class}",
+			RTarget: "green",
+			Operand: "==",
+		})
+
+	tg2 := tg1.Copy()
+	tg2.Name = "web2"
+	tg2.Constraints[0].RTarget = "blue"
+	job.TaskGroups = append(job.TaskGroups, tg2)
+	noErr(t, h.State.UpsertJob(h.NextIndex(), job))
+
+	// Create an allocation on each node
+	alloc := mock.Alloc()
+	alloc.Job = job
+	alloc.JobID = job.ID
+	alloc.NodeID = node.ID
+	alloc.Name = "my-job.web[0]"
+	alloc.TaskGroup = "web"
+
+	alloc2 := mock.Alloc()
+	alloc2.Job = job
+	alloc2.JobID = job.ID
+	alloc2.NodeID = node2.ID
+	alloc2.Name = "my-job.web2[0]"
+	alloc2.TaskGroup = "web2"
+	noErr(t, h.State.UpsertAllocs(h.NextIndex(), []*structs.Allocation{alloc, alloc2}))
+
+	// Create a mock evaluation to deal with drain
+	eval := &structs.Evaluation{
+		ID:          structs.GenerateUUID(),
+		Priority:    50,
+		TriggeredBy: structs.EvalTriggerNodeUpdate,
+		JobID:       job.ID,
+		NodeID:      node.ID,
+	}
+
+	// Process the evaluation
+	err := h.Process(NewSystemScheduler, eval)
+	if err != nil {
+		t.Fatalf("err: %v", err)
+	}
+
+	// Ensure a single plan
+	if len(h.Plans) != 1 {
+		t.Fatalf("bad: %#v", h.Plans)
+	}
+	plan := h.Plans[0]
+
+	// Ensure the plan evicted the alloc on the failed node
+	planned := plan.NodeUpdate[node.ID]
+	if len(planned) != 1 {
+		t.Fatalf("bad: %#v", plan)
+	}
+
+	// Ensure the plan didn't place
+	if len(plan.NodeAllocation) != 0 {
+		t.Fatalf("bad: %#v", plan)
+	}
+
+	// Ensure the allocations is stopped
+	if planned[0].DesiredStatus != structs.AllocDesiredStatusStop {
+		t.Fatalf("bad: %#v", planned[0])
+	}
+
+	h.AssertEvalStatus(t, structs.EvalStatusComplete)
 }